Remove handleChildResult in favor of a channel
We needed to have a way to do a select between the notifyCh and the terminateCh
on the Client algorithm, and it was not possible to do a "select case" using a
function.

Now, whenever a Child goroutine finishes it's work, it may report to either the
terminateCh or the notifyCh (whichever reads first), both of them are read by
the Supervisor, but in different execution states, terminateCh is read when
terminating all children, notifyCh is read when executing normal supervisor
restart operations.