Allow the creation of ChildSpec at start time
Instead of passing an static list of ChildSpec to the SupervisorSpec definition,
we can pass a function that returns the children nodes.

This is very powerful as it allows us to bound resources per sub-tree start. For
example, imagine an scenario where you have a group of worker goroutines relying
on a gochan for communication; with this change, the initialization of the
gochan can be done at c.ChildSpec construction time, allowing the creation of a
new gochan when the supervisor gets restarted.